test code in iotcore
