--- src/main/java/ca/uwaterloo/drinkmasterapi/common/exception/BaseException.java
@@ -1,19 +1,0 @@
-package ca.uwaterloo.drinkmasterapi.common.exception;
-
-public abstract class BaseException extends Exception {
-
-    public BaseException() { }
-
-    public BaseException(String message) {
-        super(message);
-    }
-
-    public BaseException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public BaseException(Throwable cause) {
-        super(cause);
-    }
-
-}
