--- src/main/java/ca/uwaterloo/drinkmasterapi/config/CORS.java
@@ -1,15 +1,0 @@
-package ca.uwaterloo.drinkmasterapi.config;
-
-import org.springframework.stereotype.Component;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-
-@Component
-public class CORS implements WebMvcConfigurer {
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/*")
-                .allowedMethods("PUT, GET, POST, DELETE, OPTIONS, PATCH");
-    }
-}
