+++ src/main/java/ca/uwaterloo/drinkmasterapi/feature/mqtt/controller/MqttClientController.java
@@ -1,13 +1,10 @@
 package ca.uwaterloo.drinkmasterapi.feature.mqtt.controller;
 
-import ca.uwaterloo.drinkmasterapi.feature.mqtt.service.MqttClientService;
+import ca.uwaterloo.drinkmasterapi.feature.mqtt.service.IMqttClientService;
 import com.amazonaws.services.iot.client.AWSIotException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import java.io.UnsupportedEncodingException;
-import java.security.NoSuchAlgorithmException;
 
 @RestController
 @CrossOrigin
@@ -15,29 +12,7 @@ import java.security.NoSuchAlgorithmException;
 public class MqttClientController {
 
     @Autowired
-    private MqttClientService mqttClientServiceImpl;
-
-    @GetMapping("/gdpr")
-    public ResponseEntity publishGdprMessage(@RequestParam("locationId") final String locationId) throws AWSIotException {
-        this.mqttClientServiceImpl.publishGdprMessage(locationId);
-        return ResponseEntity.ok().build();
-    }
-
-    @GetMapping("/ota")
-    public ResponseEntity publishOtaMessage(@RequestParam("locationId") final String locationId,
-                                            @RequestParam("url") final String url,
-                                            @RequestParam("productIndex") final String productIndex) throws AWSIotException, UnsupportedEncodingException, NoSuchAlgorithmException {
-        this.mqttClientServiceImpl.publishOtaMessage(locationId, url, productIndex);
-        return ResponseEntity.ok().build();
-    }
-
-    @GetMapping("/label")
-    public ResponseEntity publishLabelMessage(@RequestParam("locationId") final String locationId,
-                                              @RequestParam("baseUrl") final String baseUrl,
-                                              @RequestParam("productIndex") final String productIndex) throws AWSIotException {
-        this.mqttClientServiceImpl.publishLabelMessage(locationId, baseUrl, productIndex);
-        return ResponseEntity.ok().build();
-    }
+    private IMqttClientService iMqttClientServiceImpl;
 
     //todo: add drinkmaster endpoint
     @GetMapping("/pour")
@@ -47,7 +22,7 @@ public class MqttClientController {
                                              @RequestParam(value = "time", required = false) final String time,
                                              @RequestParam(value = "content") final String content
     ) throws AWSIotException {
-        this.mqttClientServiceImpl.publishPourMessage(id, machineId, transId, time, content);
+        this.iMqttClientServiceImpl.publishPourMessage(id, machineId, transId, time, content);
         return ResponseEntity.ok().build();
     }
 }
