+++ src/main/resources/application.properties
@@ -4,6 +4,8 @@ server.port=8498
 #AWS IOT
 awsiot.clientEndpoint=a3dl717jrbjpot-ats.iot.us-east-1.amazonaws.com
 awsiot.clientId=drinkmaster-client
-awsiot.accessKey=AKIAWK6XIIZ45O32G75Q
-awsiot.secretKey=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
+#awsiot.accessKey=AKIAWK6XIIZ45O32G75Q
+awsiot.accessKey=AKIAWK6XIIZ4U4ZC7QGF
+#awsiot.secretKey=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
+awsiot.secretKey=ftGDhS59+eERsce3t2J9m5RFij+0XzTtOZkj1AT3
 awsiot.region=us-east-1
